Fix slight errors

+++ src/models/event.py
@@ -1,13 +1,11 @@
-from typing import Optional
-from datetime import date
 from ..services.database import Base
-from sqlalchemy import Column, Integer, String, DateTime, ForeignKey, text, BigInteger
+from sqlalchemy import Column, Integer, String, DateTime
 
 
 class Event(Base):
     __tablename__ = "Event"
     
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     name = Column(String)
     description = Column(String)
     eventTypeId = Column(Integer)
